Deprecated update(int), formatting
