+++ src/main/java/com/github/manolo8/darkbot/backpage/DispatchManager.java
@@ -41,6 +41,11 @@ public class DispatchManager {
         return data;
     }
 
+    @Deprecated
+    public boolean update(int expiryTime) {
+        return this.update((long) expiryTime);
+    }
+
     public boolean update(long expiryTime) {
         try {
             if (System.currentTimeMillis() <= lastDispatcherUpdate + expiryTime) return false;
