+++ src/main/java/com/github/manolo8/darkbot/backpage/nova/Agent.java
@@ -122,169 +122,5 @@ public class Agent {
         this.desc = desc;
     }
 
-    public class DismissCost {
-        private String item;
-        private int amount;
 
-        public String getItem() {
-            return item;
-        }
-
-        public void setItem(String item) {
-            this.item = item;
-        }
-
-        public int getAmount() {
-            return amount;
-        }
-
-        public void setAmount(int amount) {
-            this.amount = amount;
-        }
-    }
-
-    public class Perk {
-        private int perkId;
-        private String perkTypeId;
-        private int level;
-        private int levelMax;
-        private int drain;
-        private double modifier;
-        private String name;
-        private String desc;
-        private String drainDesc;
-        private List<Upgrade> upgrades = new ArrayList<Upgrade>();
-
-        public int getPerkId() {
-            return perkId;
-        }
-
-        public void setPerkId(int perkId) {
-            this.perkId = perkId;
-        }
-
-        public String getPerkTypeId() {
-            return perkTypeId;
-        }
-
-        public void setPerkTypeId(String perkTypeId) {
-            this.perkTypeId = perkTypeId;
-        }
-
-        public int getLevel() {
-            return level;
-        }
-
-        public void setLevel(int level) {
-            this.level = level;
-        }
-
-        public int getLevelMax() {
-            return levelMax;
-        }
-
-        public void setLevelMax(int levelMax) {
-            this.levelMax = levelMax;
-        }
-
-        public int getDrain() {
-            return drain;
-        }
-
-        public void setDrain(int drain) {
-            this.drain = drain;
-        }
-
-        public double getModifier() {
-            return modifier;
-        }
-
-        public void setModifier(double modifier) {
-            this.modifier = modifier;
-        }
-
-        public String getName() {
-            return name;
-        }
-
-        public void setName(String name) {
-            this.name = name;
-        }
-
-        public String getDesc() {
-            return desc;
-        }
-
-        public void setDesc(String desc) {
-            this.desc = desc;
-        }
-
-        public String getDrainDesc() {
-            return drainDesc;
-        }
-
-        public void setDrainDesc(String drainDesc) {
-            this.drainDesc = drainDesc;
-        }
-
-        public List<Upgrade> getUpgrades() {
-            return upgrades;
-        }
-
-        public void setUpgrades(List<Upgrade> upgrades) {
-            this.upgrades = upgrades;
-        }
-
-    }
-
-    public class Cost {
-        private int amount;
-        private String item;
-
-        public int getAmount() {
-            return amount;
-        }
-
-        public void setAmount(int amount) {
-            this.amount = amount;
-        }
-
-        public String getItem() {
-            return item;
-        }
-
-        public void setItem(String item) {
-            this.item = item;
-        }
-    }
-
-    public class Upgrade {
-        private int level;
-        private double modifier;
-        private List<Cost> cost = new ArrayList<Cost>();
-
-        public int getLevel() {
-            return level;
-        }
-
-        public void setLevel(int level) {
-            this.level = level;
-        }
-
-        public double getModifier() {
-            return modifier;
-        }
-
-        public void setModifier(double modifier) {
-            this.modifier = modifier;
-        }
-
-        public List<Cost> getCost() {
-            return cost;
-        }
-
-        public void setCost(List<Cost> cost) {
-            this.cost = cost;
-        }
-    }
 }
