+++ src/main/java/com/github/manolo8/darkbot/backpage/nova/Cost.java
@@ -1,0 +1,22 @@
+package com.github.manolo8.darkbot.backpage.nova;
+
+public class Cost {
+    private int amount;
+    private String item;
+
+    public int getAmount() {
+        return amount;
+    }
+
+    public void setAmount(int amount) {
+        this.amount = amount;
+    }
+
+    public String getItem() {
+        return item;
+    }
+
+    public void setItem(String item) {
+        this.item = item;
+    }
+}
