+++ src/main/java/com/github/manolo8/darkbot/backpage/nova/DismissCost.java
@@ -1,0 +1,22 @@
+package com.github.manolo8.darkbot.backpage.nova;
+
+public class DismissCost {
+    private String item;
+    private int amount;
+
+    public String getItem() {
+        return item;
+    }
+
+    public void setItem(String item) {
+        this.item = item;
+    }
+
+    public int getAmount() {
+        return amount;
+    }
+
+    public void setAmount(int amount) {
+        this.amount = amount;
+    }
+}
